Soft delete users and posts

+++ src/main/java/app/fyreplace/api/endpoints/PostsEndpoint.java
@@ -65,16 +65,18 @@ public final class PostsEndpoint {
         final var stream =
                 switch (type) {
                     case SUBSCRIBED_TO -> Subscription.<Subscription>find(
-                                    "user",
+                                    "user = ?1 and post.deleted = false",
                                     Sort.by("post.dateCreated", "post.id").direction(direction),
                                     user)
                             .page(page, pagingSize)
                             .stream()
                             .map(s -> s.post);
                     case PUBLISHED -> Post.<Post>find("author = ?1 and published = true", basicSort, user)
+                            .filter("existing")
                             .page(page, pagingSize)
                             .stream();
                     case DRAFTS -> Post.<Post>find("author = ?1 and published = false", basicSort, user)
+                            .filter("existing")
                             .page(page, pagingSize)
                             .stream();
                 };
@@ -123,7 +125,13 @@ public final class PostsEndpoint {
         final var user = User.getFromSecurityContext(context);
         final var post = Post.<Post>findById(id);
         Post.validateAccess(post, user, null, true);
-        post.delete();
+
+        if (post.published) {
+            post.softDelete();
+        } else {
+            post.delete();
+        }
+
         return Response.noContent().build();
     }
 
@@ -226,9 +234,9 @@ public final class PostsEndpoint {
     public long count(@QueryParam("type") @NotNull final PostListingType type) {
         final var user = User.getFromSecurityContext(context);
         return switch (type) {
-            case SUBSCRIBED_TO -> Subscription.count("user", user);
-            case PUBLISHED -> Post.count("author = ?1 and published = true", user);
-            case DRAFTS -> Post.count("author = ?1 and published = false", user);
+            case SUBSCRIBED_TO -> Subscription.count("user = ?1 and post.deleted = false", user);
+            case PUBLISHED -> Post.count("author = ?1 and published = true and deleted = false", user);
+            case DRAFTS -> Post.count("author = ?1 and published = false and deleted = false", user);
         };
     }
 
@@ -240,13 +248,19 @@ public final class PostsEndpoint {
         final var user = User.getFromSecurityContext(context);
 
         try (final var stream = Post.<Post>find(
-                "author != ?1 and dateCreated > ?2 and published = true and life > 0"
-                        + "and id not in (select post.id from Vote where user = ?1)"
-                        + "and author.id not in (select target.id from Block where source = ?1)"
-                        + "and author.id not in (select source.id from Block where target = ?1)",
+                """
+                author != ?1
+                and dateCreated > ?2
+                and published = true
+                and life > 0
+                and id not in (select post.id from Vote where user = ?1)
+                and author.id not in (select target.id from Block where source = ?1)
+                and author.id not in (select source.id from Block where target = ?1)
+                """,
                 Sort.by("life", "dateCreated", "id"),
                 user,
                 Instant.now().minus(Post.shelfLife))
+                .filter("existing")
                 .range(0, 2)
                 .stream()) {
             return stream.peek(p -> p.setCurrentUser(user)).toList();
